fix:코드 깔끔하게 만들기
