+++ src/prisma/prisma.service.ts
@@ -1,20 +1,11 @@
-import {
-  INestApplication,
-  Inject,
-  Injectable,
-  OnModuleInit,
-  forwardRef,
-} from '@nestjs/common';
+import { INestApplication, Injectable, OnModuleInit } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client';
 import settings from '../settings';
-import { ReviewMiddleware } from './middleware/prisma.reviews';
+import { mediator } from './middleware/mediator';
 
 @Injectable()
 export class PrismaService extends PrismaClient implements OnModuleInit {
-  constructor(
-    @Inject(forwardRef(() => ReviewMiddleware))
-    private readonly reviewMiddleware: ReviewMiddleware,
-  ) {
+  constructor() {
     const ormOption = settings().ormconfig();
     super(ormOption);
   }
@@ -27,42 +18,14 @@ export class PrismaService extends PrismaClient implements OnModuleInit {
       // console.log(`Query: ${e.query} ${e.params}`);
     });
     this.$use(async (params, next) => {
-      if (params.model === 'review_review') {
-        //todo: determine reuslt type
-        if (
-          params.action === 'create' ||
-          params.action === 'update' ||
-          params.action === 'upsert'
-        ) {
-          const result = await next(params);
-          await this.reviewMiddleware.reviewSavedMiddleware(
-            result,
-            params.action,
-          );
-          return result;
-        } else if (params.action === 'delete') {
-          const result = await next(params);
-          await this.reviewMiddleware.reviewDeletedMiddleware(result);
-          return result;
-        }
-      } else if (params.model === 'review_reviewvote') {
-        if (
-          params.action === 'create' ||
-          params.action === 'update' ||
-          params.action === 'upsert'
-        ) {
-          const result = await next(params);
-          await this.reviewMiddleware.reviewVoteSavedMiddleware(result);
-          return result;
-        } else if (params.action === 'delete') {
-          const result = await next(params);
-          await this.reviewMiddleware.reviewVoteDeletedMiddleware(result);
-          return result;
-        }
+      const signal = mediator(this, params);
+      if (signal) {
+        const result = await next(params);
+        await signal.execute(params, result);
+        return result;
       }
       return next(params);
     });
-    //this.reviewMiddleware.reviewSavedMiddleware();
   }
 
   async enableShutdownHooks(app: INestApplication) {
